Bug and code smell fixes

+++ src/main/java/com/configcat/DateTimeUtils.java
@@ -26,6 +26,6 @@ public class DateTimeUtils {
     }
 
     public static double getUnixSeconds(Date date) {
-        return Long.valueOf(date.getTime()).doubleValue() / 1000;
+        return date.getTime() / 1000d;
     }
 }
